problem: the kube-dash ui is not well known among the engineers who might benefit from it the most
solution: add a simple command - `kube-dash` - to smooth the path to
accessing kubernetes-dashboard web ui.

I love the cli but I was reminded from a talk at google next 2017 that
not all engineers are. We still want these engineers to be able to be
productive with kubernetes and hopefully this will help them discover
and use the web dashboard.